Dongha

+++ src/main/java/com/example/greenlast/dto/ClassDTO.java
@@ -13,7 +13,7 @@ public class ClassDTO {
     private String classCapsule;
     private String classSubtitle;
     private String description;
-    private double classPrice;
+    private int classPrice;
     private String classCategory;
     private String classCreateDate;
     private String classTerm;
